+++ lib/ui/screens/add_card_screen.dart
@@ -1,6 +1,6 @@
 import 'package:figma_shopping_app/generated/assets.dart';
-import 'package:figma_shopping_app/screens/onboarding_screen_scaffold.dart';
-import 'package:figma_shopping_app/screens/order_confirmed_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding_screen_scaffold.dart';
+import 'package:figma_shopping_app/ui/screens/order_confirmed_screen.dart';
 import 'package:flutter/material.dart';
 
 class AddCardScreen extends StatefulWidget {
