+++ lib/ui/screens/dashboard_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class DashboardScreen extends StatelessWidget {
+  const DashboardScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Dashboard Screen"),
+      ),
+    );
+  }
+}
