+++ lib/ui/screens/onboarding/forgot_pwd_screen.dart
@@ -1,7 +1,7 @@
 import 'package:figma_shopping_app/generated/assets.dart';
-import 'package:figma_shopping_app/screens/onboarding/signup_onboarding_screen.dart';
-import 'package:figma_shopping_app/screens/onboarding/verification_code_screen.dart';
-import 'package:figma_shopping_app/screens/onboarding_screen_scaffold.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding/signup_onboarding_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding/verification_code_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding_screen_scaffold.dart';
 import 'package:flutter/material.dart';
 
 class ForgotPwdScreen extends StatelessWidget {
