+++ lib/ui/screens/onboarding/login_screen.dart
@@ -1,23 +1,36 @@
 import 'package:awesome_extensions/awesome_extensions.dart';
-import 'package:figma_shopping_app/colors.dart';
 import 'package:figma_shopping_app/generated/assets.dart';
-import 'package:figma_shopping_app/screens/onboarding_screen_scaffold.dart';
+import 'package:figma_shopping_app/ui/screens/gender_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding/forgot_pwd_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding/signup_onboarding_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding_screen_scaffold.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter_svg/svg.dart';
 
-class SignupOnboardingScreen extends StatelessWidget {
-  const SignupOnboardingScreen({super.key});
+class LoginScreen extends StatelessWidget {
+  const LoginScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return OnboardingScreenScaffold(
-      title: "Sign Up",
-      buttonText: "Sign Up",
-      onButtonClick: () {},
+      title: "Welcome",
+      subtitle: "Please Enter your data to Continue",
+      lowerText:
+          "By Connecting your Account you confirm that you Agree with our",
+      lowerTextAction: "Terms and Conditions",
+      buttonText: "Login",
+      onButtonClick: () {
+        Navigator.of(context).push(
+          MaterialPageRoute(
+            builder: (context) => const GenderScreen(),
+          ),
+        );
+      },
       child: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 20),
         child: Column(
+          mainAxisSize: MainAxisSize.min,
           children: [
             TextField(
               decoration: InputDecoration(
@@ -48,17 +61,24 @@ class SignupOnboardingScreen extends StatelessWidget {
                 ),
               ),
             ),
-            TextField(
-              decoration: InputDecoration(
-                labelText: "Email Address",
-                labelStyle: labelTextStyle,
-                enabledBorder: textFieldBorder,
-                suffixIconConstraints: const BoxConstraints(),
-                suffixIcon: SvgPicture.asset(
-                  Assets.assetsCheckIcon,
-                  alignment: Alignment.bottomCenter,
-                  width: 10,
-                  height: 10,
+            Padding(
+              padding: const EdgeInsets.only(top: 30),
+              child: Align(
+                alignment: Alignment.centerRight,
+                child: InkWell(
+                  onTap: () {
+                    Navigator.of(context).push(
+                      MaterialPageRoute(
+                          builder: (context) => const ForgotPwdScreen()),
+                    );
+                  },
+                  child: const Text(
+                    "Forgot Password?",
+                    style: TextStyle(
+                      color: Color(0xFFEA4335),
+                      fontSize: 15,
+                    ),
+                  ),
                 ),
               ),
             ),
@@ -94,12 +114,3 @@ class SignupOnboardingScreen extends StatelessWidget {
     );
   }
 }
-
-const labelTextStyle = TextStyle(
-  color: subtitleColor,
-  fontSize: 16,
-);
-
-const textFieldBorder = UnderlineInputBorder(
-  borderSide: BorderSide(color: textFieldUnderlineColor),
-);
