+++ lib/ui/screens/onboarding/new_pwd_screen.dart
@@ -1,6 +1,6 @@
-import 'package:figma_shopping_app/screens/cart_screen.dart';
-import 'package:figma_shopping_app/screens/onboarding/signup_onboarding_screen.dart';
-import 'package:figma_shopping_app/screens/onboarding_screen_scaffold.dart';
+import 'package:figma_shopping_app/ui/screens/cart_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding/signup_onboarding_screen.dart';
+import 'package:figma_shopping_app/ui/screens/onboarding_screen_scaffold.dart';
 import 'package:flutter/material.dart';
 
 class NewPwdScreen extends StatelessWidget {
